chore(marketplace): remove orphaned skill-browser component
The standalone SkillBrowser surface was retired from the Marketplace hub
(cave-vewe) but the 944-line component, its test, and ~609 lines of
skill-browser__* CSS stayed behind. Nothing renders <SkillBrowser>;
consumers only imported the SkillBrowserEntry type from it.

- Move SkillBrowserEntry to src/lib/skill-directory.ts (its canonical
  home next to the shared directory helpers) and rewire all importers.
- Delete src/components/skill-browser.tsx + skill-browser.test.ts; the
  live Marketplace-hub wiring assertions it carried (Explore cards,
  drawer, keying, debounced registry search, retired SkillCard/skills
  tabpanel) move to marketplace/skill-explore.test.ts, registered in
  scripts/run-tests.mjs.
- Drop the skill-builder.test.ts read of the deleted file and its
  empty-state CTA assertion; repoint the roles-tools-navigation pin at
  the type's new import path.
- Prune the unused .skill-browser__* rules from surface-marketplace.css
  and bank the drift-ratchet progress (font 154->151, space 1565->1526,
  radius 211->208, inline 231->230).
