+++ scripts/run-tests.mjs
@@ -309,6 +309,7 @@ export const SUITES = {
     "src/components/marketplace/craft-create-drawer.test.ts",
     "src/lib/craft-agent-prompt.test.ts",
     "src/components/marketplace/skill-builder.test.ts",
+    "src/components/marketplace/skill-explore.test.ts",
     "src/lib/skill-templates.test.ts",
     "src/lib/skill-draft.test.ts",
     "src/lib/skill-build-format.test.ts",
@@ -317,7 +318,6 @@ export const SUITES = {
     "src/lib/skill-agent-prompt.test.ts",
     "src/lib/skill-directory.test.ts",
     "src/components/terminal-scope.test.ts",
-    "src/components/skill-browser.test.ts",
     "src/components/top-bar.test.ts",
     "src/components/command-palette.test.ts",
     "src/components/command-palette-a11y.test.ts",
