+++ src/lib/design-token-drift.test.ts
@@ -42,11 +42,11 @@ import {
 // Current counts as of the P3 codemod PR. Only lower these (banking progress)
 // or raise them with an explicit justification in your PR.
 const BASELINES = {
-  offScaleFontSizePx: 155, // 10.5px/11.5px/… — need per-case renormalization to the type scale (banked: canvas-editor.css on the type scale; familiar-tab retokenized the old hero/section styles; research-desk tabs shipped fully on the type scale, -3; familiar-tab five-section rebuild shipped fully on the type scale, -1; projects access page rebuild shipped fully on the type scale, -2). +1: Research Reader (Research Reader.dc.html handoff) — the EB Garamond findings hero <h1> keeps the design's 31px display size (between --text-display 28px and the next step); every other reader font-size uses the type scale and compact mono labels use the `font:` shorthand.
-  offScaleSpacingPx: 1576, // off-4px-grid pad/margin/gap components — banked: canvas-editor.css on the space scale (-1) + familiar-tab snapped to the 4px grid (-5). +123: research-desk redesign (cave-dl74) — the five tab sheets keep the design's compact even 2px-offset rhythm (6/10/14/18/22px gaps and paddings, same family the surface it replaced used); odd values were snapped to the grid in the same PR. +7: home-minimal PR (2026-07-22) — new home-composer/landing-composer.css + hearth-continuations.css compact UI rhythms (7px gaps, 6px accent-dot, 5px pill gap, 13px/14px organic pill padding — off-scale by design for visual compactness). -10: projects access page rebuild (cave-luq4) replaced the hub CSS fully on the space scale; re-banked at 1467 after rebasing onto main commits 7cf9d13f0..fbcadfd3b which added +4 (still -6 vs the pre-rebuild 1473). -1: drive-by banked in the hex-token chunk PR (cave-gyh2). +26: Queue tab redesign (Queue.dc.html handoff) — the refreshed familiar-work-queue.css keeps the design's compact rhythm (7px control gaps, 6px row-meta gaps, 5px check-dot gaps, 3px accent rail, 30px action-button height, 130px note-textarea min-height, 180px forward-menu min-width, 37px menu offset); on-scale values were tokenized by tokenize-css.mjs in the same PR and font-size literals were snapped to the type scale. -2: Tasks toolbar redesign (tasks-list-redesign-refresh) swapped the compact board header for the design's roomier command bar; net off-scale spacing dropped even though the new toolbar adds off-scale rhythm, because the removed compact-header + search-wrap block carried more off-grid values. +40: launcher 3a home dashboard (home-dashboard.css) — the work-led dashboard shell keeps the design mock's compact rhythm (13/22px chrome & board padding, 9/10/11/18/26px rail & row gaps/padding, 2/3/5/6/7px micro-gaps incl. the Resume CTA pill), off the 4px grid by design; the redesign also RECLAIMED the retired hearth/hero/Continue sheets, so this is the NET delta after removal; on-grid values were tokenized by the codemod in the same PR. +20: Review Deck redesign (Review Deck.dc.html handoff) — new review-deck.css keeps the design's compact rhythm (2px stat/pill micro-gaps, 3px pill padding, 5px pill gaps, 6px row/eyebrow margins, 7px key/button gaps, 14px panel-head padding), off the 4px grid by design; the trivially-equivalent 9/7/13px values were snapped to the space scale in the same PR and font-size literals to the type scale. +34: Familiar Analytics modernization (cave-q8uc, Familiar Analytics.dc.html handoff) — the refreshed familiar-analytics surface (surface-reporting.css) keeps the design mock's compact rhythm: 14px heal/contract-detail card padding, 10px attention/heal gaps, 11/13px collapse & button paddings, 7/9px heal-card button padding, 5/6px micro-gaps, 2/3px chip nudges — off the 4px grid by design for the dense card/chip/modal layout; on-scale values were tokenized by tokenize-css.mjs in the same PR. +3: Marketplace "Explore" redesign (Marketplace.dc.html handoff) — the merged Explore rail + toolbar keep the mock's compact rhythm (1px rail-row gap, 2px view-toggle inner padding, 10px rail-label inline padding); on-scale values were tokenized by tokenize-css.mjs in the same PR. -10: hearth-home restore / dashboard-to-new-chat — the hearth sheets' +7 compact rhythm returned, but relocating the dashboard into the chat empty state retired its chrome/dock/project-button blocks (13/22px chrome + 14/22px dock padding, 9/10/11px project paddings), a net reclaim. -12: new-chat dashboard simplification (cave-gxap) — the context rail retired (18px rail padding/gap, 9/10/11px project/quick/pick paddings & gaps, 2px micro-margin) and the board's 26px padding/gap snapped to the space scale; the rail-less board is a capped no-scroll column. +11: Research Reader (Research Reader.dc.html handoff) — the reader chrome keeps the design's compact reading rhythm off the 4px grid (34px page inset & doc top padding, 48px reading-column gutter, 22/14px contents-rail padding, 7px toc/link paddings, 15px lede indent, 20px list indent, 18px rail padding, and 5/6/7/9/10/11px micro paddings & gaps across the toolbar, source cards, and focus overlay); trivially on-grid values were tokenized by tokenize-css.mjs in the same PR.
-  offScaleRadiusPx: 212, // 4px/6px/10px/14px/… radii between the sanctioned steps. +5: research-desk 2px/4px accent-mark radii (short solid marks, not container corners). -5: projects access page rebuild removed the hub's off-step radii. +1: Research Reader accent-mark radius (the section-heading 2px tick, same short-solid-mark family as the research desk).
+  offScaleFontSizePx: 152, // 10.5px/11.5px/… — need per-case renormalization to the type scale (banked: canvas-editor.css on the type scale; familiar-tab retokenized the old hero/section styles; research-desk tabs shipped fully on the type scale, -3; familiar-tab five-section rebuild shipped fully on the type scale, -1; projects access page rebuild shipped fully on the type scale, -2). +1: Research Reader (Research Reader.dc.html handoff) — the EB Garamond findings hero <h1> keeps the design's 31px display size (between --text-display 28px and the next step); every other reader font-size uses the type scale and compact mono labels use the `font:` shorthand. -3: orphaned skill-browser removal reclaimed its off-scale sizes.
+  offScaleSpacingPx: 1537, // off-4px-grid pad/margin/gap components — banked: canvas-editor.css on the space scale (-1) + familiar-tab snapped to the 4px grid (-5). +123: research-desk redesign (cave-dl74) — the five tab sheets keep the design's compact even 2px-offset rhythm (6/10/14/18/22px gaps and paddings, same family the surface it replaced used); odd values were snapped to the grid in the same PR. +7: home-minimal PR (2026-07-22) — new home-composer/landing-composer.css + hearth-continuations.css compact UI rhythms (7px gaps, 6px accent-dot, 5px pill gap, 13px/14px organic pill padding — off-scale by design for visual compactness). -10: projects access page rebuild (cave-luq4) replaced the hub CSS fully on the space scale; re-banked at 1467 after rebasing onto main commits 7cf9d13f0..fbcadfd3b which added +4 (still -6 vs the pre-rebuild 1473). -1: drive-by banked in the hex-token chunk PR (cave-gyh2). +26: Queue tab redesign (Queue.dc.html handoff) — the refreshed familiar-work-queue.css keeps the design's compact rhythm (7px control gaps, 6px row-meta gaps, 5px check-dot gaps, 3px accent rail, 30px action-button height, 130px note-textarea min-height, 180px forward-menu min-width, 37px menu offset); on-scale values were tokenized by tokenize-css.mjs in the same PR and font-size literals were snapped to the type scale. -2: Tasks toolbar redesign (tasks-list-redesign-refresh) swapped the compact board header for the design's roomier command bar; net off-scale spacing dropped even though the new toolbar adds off-scale rhythm, because the removed compact-header + search-wrap block carried more off-grid values. +40: launcher 3a home dashboard (home-dashboard.css) — the work-led dashboard shell keeps the design mock's compact rhythm (13/22px chrome & board padding, 9/10/11/18/26px rail & row gaps/padding, 2/3/5/6/7px micro-gaps incl. the Resume CTA pill), off the 4px grid by design; the redesign also RECLAIMED the retired hearth/hero/Continue sheets, so this is the NET delta after removal; on-grid values were tokenized by the codemod in the same PR. +20: Review Deck redesign (Review Deck.dc.html handoff) — new review-deck.css keeps the design's compact rhythm (2px stat/pill micro-gaps, 3px pill padding, 5px pill gaps, 6px row/eyebrow margins, 7px key/button gaps, 14px panel-head padding), off the 4px grid by design; the trivially-equivalent 9/7/13px values were snapped to the space scale in the same PR and font-size literals to the type scale. +34: Familiar Analytics modernization (cave-q8uc, Familiar Analytics.dc.html handoff) — the refreshed familiar-analytics surface (surface-reporting.css) keeps the design mock's compact rhythm: 14px heal/contract-detail card padding, 10px attention/heal gaps, 11/13px collapse & button paddings, 7/9px heal-card button padding, 5/6px micro-gaps, 2/3px chip nudges — off the 4px grid by design for the dense card/chip/modal layout; on-scale values were tokenized by tokenize-css.mjs in the same PR. +3: Marketplace "Explore" redesign (Marketplace.dc.html handoff) — the merged Explore rail + toolbar keep the mock's compact rhythm (1px rail-row gap, 2px view-toggle inner padding, 10px rail-label inline padding); on-scale values were tokenized by tokenize-css.mjs in the same PR. -10: hearth-home restore / dashboard-to-new-chat — the hearth sheets' +7 compact rhythm returned, but relocating the dashboard into the chat empty state retired its chrome/dock/project-button blocks (13/22px chrome + 14/22px dock padding, 9/10/11px project paddings), a net reclaim. -12: new-chat dashboard simplification (cave-gxap) — the context rail retired (18px rail padding/gap, 9/10/11px project/quick/pick paddings & gaps, 2px micro-margin) and the board's 26px padding/gap snapped to the space scale; the rail-less board is a capped no-scroll column. +11: Research Reader (Research Reader.dc.html handoff) — the reader chrome keeps the design's compact reading rhythm off the 4px grid (34px page inset & doc top padding, 48px reading-column gutter, 22/14px contents-rail padding, 7px toc/link paddings, 15px lede indent, 20px list indent, 18px rail padding, and 5/6/7/9/10/11px micro paddings & gaps across the toolbar, source cards, and focus overlay); trivially on-grid values were tokenized by tokenize-css.mjs in the same PR. -39: orphaned skill-browser removal (cave-vewe follow-up) — deleting the retired surface's ~609-line .skill-browser__* block reclaimed its compact off-grid rhythm wholesale.
+  offScaleRadiusPx: 209, // 4px/6px/10px/14px/… radii between the sanctioned steps. +5: research-desk 2px/4px accent-mark radii (short solid marks, not container corners). -5: projects access page rebuild removed the hub's off-step radii. +1: Research Reader accent-mark radius (the section-heading 2px tick, same short-solid-mark family as the research desk). -3: orphaned skill-browser removal reclaimed its off-step radii.
   hexOutsideDefinitions: 0, // hex in render CSS (token definitions excluded) — -104: cave-gyh2 chunk 1 (dropped stale var(--token, #hex) fallbacks; mapped accent/danger/success fills to their semantic foregrounds; promoted --color-success-foreground and the codex --cv-* strays). -1 banked in the design-doc reconcile PR (cave-kf3x). -51: cave-yxiz chunk 2 zeroed the ratchet — document grounds (sketch/preview/thumb/QR) now share the fixed --surface-paper token in foundations.css; GitHub state badges promote --gh-merged/--gh-merged-ink and pair open/closed with the semantic status foregrounds; profile-card strays joined the --pfc-* palette; the magic-cast spell art promotes --spell-violet/--spell-core; QR ink, dashboard mark ink, and the avatar photo-overlay ink became local definitions; and pure shade/alpha arithmetic in color-mix()/mask gradients uses the CSS black/white keywords (sanctioned: they're mix anchors, not colors — matching the pre-existing keyword usage in dashboard.css and surface-compact-calendar.css). New hexes belong in token definitions; keywords are only for mix/mask arithmetic.
-  inlineTsxStyles: 231, // style={{…}} in TSX; many are legit dynamic values (banked from 500 — the count had drifted far below the old ceiling; -2: projects access page rebuild dropped the hub's inline styles; +1: vendored Blaze.tsx canvas host uses dynamic style spread for WebGL positioning; +1: vendored Peel.tsx canvas host for page-peel reveal effect uses runtime-mutated inline styles; +2: project-setup modal color swatches paint runtime-only values (per-root tint + oklch palette choice))
+  inlineTsxStyles: 230, // style={{…}} in TSX; many are legit dynamic values (banked from 500 — the count had drifted far below the old ceiling; -2: projects access page rebuild dropped the hub's inline styles; +1: vendored Blaze.tsx canvas host uses dynamic style spread for WebGL positioning; +1: vendored Peel.tsx canvas host for page-peel reveal effect uses runtime-mutated inline styles; +2: project-setup modal color swatches paint runtime-only values (per-root tint + oklch palette choice); -1: orphaned skill-browser removal dropped its inline style)
 };
 
 // ── unit sanity for the codemod transform ───────────────────────────────────
